aggregator: Update include following commits
commit: 447da9a6105b66591b9c9677c63b06af75a11cac
* aggregator: handle erroneous case of the first buffer before segment
Segment event always should arrive before the first buffer, however if
that happens due to some complex pipelines manipulations, we don't want
the process to crash or go into errouneous state.
In partigular an abort gets raised by the videoaggregator in this case.
To avoid that we propose to safely assume we have received a segment in
which the first buffer is valid.

commit: 35f740b09f9d10493af1030904478ed3b035c129
* aggregator: avoid race condition after FLUSH_STOP event
We wait until all pads have buffers in gst_aggregator_wait_and_check()
but ub case of FLUSH_STOP event first_buffer was not getting true which
is one of the condition to wait in gst_aggregator_wait_and _check()
causing fail to wait

commit: dccb2f6b5c293c2ddad76c79da74cec288b85988
* aggregator: fix use-after-free in queries processing
Test included.
The problem appears when aggregator drops the query while
it's being proccessed by the klass->sink_query handler.
This can happen on FLUSH_START event. If the query is still
in the queue, it can be safely dropped, but if it's already
in the klass->sink_query() handler, then sink pad has no
choice and has to wait for the proccessing to complete.

commit: 7829a5e5e8d2cbd964c5b9add27a1dfd96161442
* Revert "aggregator: fix use-after-free in queries processing"
This reverts commit 3943b39fee9b0fb56f3f37572c96e4571b23cc86.

commit: 3943b39fee9b0fb56f3f37572c96e4571b23cc86
* aggregator: fix use-after-free in queries processing
This is a real use-after-free caught by address sanitizer.
The problem is that while the query is proccessed by the
klass->sink_event(), it might be destroyed if there will
be a state change at this moment.

commit: 1fc183726336fcc83aa5b6d2f8aa96d4f5d0ebba
* fix for crash caused by commit 6c66ede6db9952d196da0eba91146ad5fe0235fe

commit: 2eb5bf6a6d4e73bdf1e933a3c5601917d24c359b
* aggregator: fix start-time-selection=first on negative rate
When the property "start-time-selection" is set to "first", it
calculates the start time of the output from the buffer pts
(converting it to running time of the segment), but if the
rate is negative, the real start is not the pts, but the
pts + duration, because it plays from the end of the buffer
to it's start.
As a result of this bug, in the negative rate, when the
start-time-selection=first, the first frame is dropped
by the videoaggregator (reproduced on d3d11compositor).

commit: a562ab7c6fb2e337d0aabbf8a05d1dff76dfce1f
* aggregator: Warn about the right segment's format
We were checking the head segment here, not the current segment.
Part-of: <https://gitlab.freedesktop.org/gstreamer/gstreamer/-/merge_requests/3819>